+++ assets/components/widgets/multiclassStandings/multiclassStandings.jsx
@@ -108,14 +108,21 @@ UI.widgets.MulticlassStandings = React.createClass({
 	renderName: function(name, classId) {
 		var firstInitial = name.substr(0, 1).toUpperCase() + ". ";
 		var parts = name.split(' ');
-		var name = parts[parts.length-1].substr(0, 3).toUpperCase();
-
 		var divStyle = {};
+
+		if (UI.state.controllerOptions.options.showFullStandingsName.value === "true") {
+			// show full name and increase width
+			var lastNames = parts.slice(1);
+			var name = lastNames.map(item=> item.toUpperCase());
+			divStyle.width = "120px";
+		} else {
+			// Show characters of last name by default
+			var name = parts[parts.length-1].substr(0, 3).toUpperCase();
+		}
+
 		if (UI.state.controllerOptions.options.multiclass.value === "true" && UI.getClassColour(classId) != null) {
 			classColour = UI.getClassColour(classId);
-			divStyle = {
-					backgroundColor: classColour
-			};
+			divStyle.backgroundColor = classColour;
 		}
 		return <div className="name" style={divStyle}>{firstInitial}{name}</div>
 	},